# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AutoAssist IA API
  version: 1.0.0
  description: API do AutoAssist - Ecossistema automotivo com IA. Consulte os endpoints
    para chat, manutenção preditiva, FIPE, pagamentos e mais.
servers:
- url: http://localhost:5001
  description: Desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    B2BDiagnosisInput:
      type: object
      required:
      - image
      properties:
        image:
          type: string
          description: base64 puro ou DataURL (max 15MB em base64)
        image_b64:
          type: string
        pergunta:
          type: string
          maxLength: 2000
        formato:
          type: string
          enum:
          - json
          - pdf
          default: json
...
security:
- bearerAuth: []
paths:
...
  /api/b2b/diagnosis:
    post:
      summary: Diagnóstico B2B por foto
      tags:
      - B2B
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/B2BDiagnosisInput'
      responses:
        '200':
          description: Laudo gerado
        '401':
          description: API key inválida
        '429':
          description: Rate limit / cota excedida
...